[add]add ItemRepository.kt for mvvc
